requirements.txt
pywinrm
python-nmap
requests